Migrate KYC to modal, general cleanup, show verification status to user
